Use solid fills as gradient/pattern replacement
Until we get real gradient and pattern fills in the WebGL renderer, we
use a solid fill with the color magenta. At least the user sees that
there is something on the screen and he may at least wonder why it is
magenta. That's better than seeing nothing at all.