Fix issue where DisplayObject was positioned incorrectly
Consider the case where we have a `DisplayObjectContainer` named `container` with a single `Bitmap` with a `GlassPlate` defined. Now let’s say we do `container.alignPivot()`. The result will be that the container will not be in the expected position. This is because `DisplayObjectContainer` expands its bounds by iterating over each `DisplayObject`, *including* `GlassPlate`. This change skips `GlassPlate` in the bounds calculations.