Migrates to package:web and removes deprecated WebGL
Updates the code to use `package:web` instead of the deprecated `dart:web_gl`.

This change improves compatibility with modern web standards and removes reliance on deprecated features. It changes the way that WebGL contexts are accessed and managed.

It also changes the way some functions in the library are called and how parameters are passed to web API.